remove windows warnings

+++ src/MEDWrapper/V2_1/Core/MEDdatasetNumEcrire.cxx
@@ -194,8 +194,8 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
         pflsize [0] = psize*ngauss*nbdim;
         pcount  [0] = psize*ngauss*dimutil;
 #ifdef HDF_NEW_API
-        pflmem      = (med_size *) malloc (sizeof(med_size)*pcount[0]);
-        pfldsk      = (med_size *) malloc (sizeof(med_size)*pcount[0]);
+        pflmem      = (med_size *) malloc (sizeof(med_size)*(size_t)pcount[0]);
+        pfldsk      = (med_size *) malloc (sizeof(med_size)*(size_t)pcount[0]);
 #else
         pflmem      = (med_ssize *) malloc (sizeof(med_ssize)*(size_t)pcount[0]);
         pfldsk      = (med_ssize *) malloc (sizeof(med_ssize)*(size_t)pcount[0]);
@@ -220,16 +220,16 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
             }
              
 #ifdef HDF_NEW_API2
-            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, pcount[0], (const hsize_t *) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t *) pflmem ) ) <0) 
               return -1; 
               
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, pcount[0], (const hsize_t *) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t *) pfldsk ) ) <0) 
               return -1; 
 #elif defined HDF_NEW_API
-            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, pcount[0], (const hsize_t **) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t **) pflmem ) ) <0) 
               return -1; 
               
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, pcount[0], (const hsize_t **) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t **) pfldsk ) ) <0) 
               return -1; 
 #else
             if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hssize_t **) pflmem ) ) <0) 
@@ -261,16 +261,16 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
             }
             
 #ifdef HDF_NEW_API2
-            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, pcount[0], (const hsize_t *) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t *) pflmem ) ) <0) 
               return -1; 
             
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, pcount[0], (const hsize_t *) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t *) pfldsk ) ) <0) 
               return -1; 
 #elif defined HDF_NEW_API
-            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, pcount[0], (const hsize_t **) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t **) pflmem ) ) <0) 
               return -1; 
             
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, pcount[0], (const hsize_t **) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET, (size_t)pcount[0], (const hsize_t **) pfldsk ) ) <0) 
               return -1; 
 #else
             if ( (ret = H5Sselect_elements(memspace,H5S_SELECT_SET, (size_t)pcount[0], (const hssize_t **) pflmem ) ) <0) 
@@ -335,7 +335,7 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
         pflsize [0] = psize*ngauss*nbdim;
         pcount  [0] = psize*ngauss*dimutil; /* nom pas très coherent avec count !!! A revoir */      
 #ifdef HDF_NEW_API
-        pfldsk     = (med_size *) malloc(sizeof(med_size)*pcount[0]);
+        pfldsk     = (med_size *) malloc(sizeof(med_size)*(size_t)pcount[0]);
 #else
         pfldsk     = (med_ssize *) malloc(sizeof(med_ssize)*(size_t)pcount[0]);
 #endif
@@ -354,10 +354,10 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
             }
             
 #ifdef HDF_NEW_API2
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,pcount[0], (const hsize_t *) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t *) pfldsk ) ) <0) 
               return -1;
 #elif defined HDF_NEW_API
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,pcount[0], (const hsize_t **) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t **) pfldsk ) ) <0) 
               return -1;
 #else
             if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,(size_t)pcount[0], (const hssize_t **) pfldsk ) ) <0) 
@@ -379,7 +379,7 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
               return -1;
 
 #ifdef HDF_NEW_API
-            pflmem     = (med_size *) malloc (sizeof(med_size)*pcount[0]);
+            pflmem     = (med_size *) malloc (sizeof(med_size)*(size_t)pcount[0]);
 #else
             pflmem     = (med_ssize *) malloc (sizeof(med_ssize)*(size_t)pcount[0]);
 #endif
@@ -397,16 +397,16 @@ _MEDdatasetNumEcrire(med_idt pere,char *nom, med_type_champ type,
             }
              
 #ifdef HDF_NEW_API2
-            if ( (ret = H5Sselect_elements(memspace ,H5S_SELECT_SET,pcount[0], (const hsize_t *) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace ,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t *) pflmem ) ) <0) 
               return -1; 
               
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,pcount[0], (const hsize_t *) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t *) pfldsk ) ) <0) 
               return -1;
 #elif defined HDF_NEW_API
-            if ( (ret = H5Sselect_elements(memspace ,H5S_SELECT_SET,pcount[0], (const hsize_t **) pflmem ) ) <0) 
+            if ( (ret = H5Sselect_elements(memspace ,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t **) pflmem ) ) <0) 
               return -1; 
               
-            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,pcount[0], (const hsize_t **) pfldsk ) ) <0) 
+            if ( (ret = H5Sselect_elements(dataspace,H5S_SELECT_SET,(size_t)pcount[0], (const hsize_t **) pfldsk ) ) <0) 
               return -1;
 #else
             if ( (ret = H5Sselect_elements(memspace ,H5S_SELECT_SET,(size_t)pcount[0], (const hssize_t **) pflmem ) ) <0)
